test suite: simplify
