+++ setup/scripts/test/travis-integration-test.sh
@@ -66,7 +66,6 @@ travis_fold end travis_integration_install
 
 travis_fold start travis_integration_tests
 
-python tests/check_status.py -v
 python tests/search_test.py
 python tests/dedupe_test.py
