+++ travis/integration_test-script.sh
@@ -144,14 +144,6 @@ else
     fail $test
 fi
 
-test="Check Status test"
-echo "Running: $test"
-if python3 tests/check_status.py ; then
-    success $test
-else
-    fail $test
-fi
-
 test="Dedupe integration tests"
 echo "Running: $test"
 if python3 tests/dedupe_test.py ; then
@@ -160,7 +152,13 @@ else
     fail $test
 fi
 
-
+test="Check Various Pages integration test"
+echo "Running: $test"
+if python3 tests/check_various_pages.py ; then
+    success $test
+else
+    fail $test
+fi
 
 # The below tests are commented out because they are still an unstable work in progress
 ## Once Ready they can be uncommented.
@@ -170,13 +168,6 @@ fi
 #     echo "Success: Import Scanner integration tests passed" 
 # else
 #     echo "Error: Import Scanner integration test failed"; exit 1
-# fi
-
-# echo "Check Status UI integration test"
-# if python3 tests/check_status_ui.py ; then
-#     echo "Success: Check Status UI tests passed"
-# else
-#     echo "Error: Check Status UI test failed"; exit 1
 # fi
 
 # echo "Zap integration test"
